requirements.txt
django==4.1.13
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4
altair
matplotlib
seaborn
scikit-learn
joblib 
python-dotenv
